api/requirements.txt
fastapi
uvicorn[standard]
pandas
kafka-python